extract action fetching into hook
