+++ packages/web-app/src/pages/newProposal.tsx
@@ -20,12 +20,11 @@ import {useNetwork} from 'context/network';
 import {useDaoParam} from 'hooks/useDaoParam';
 import {Loading} from 'components/temporary';
 import {CreateProposalProvider} from 'context/createProposal';
-import {useDaoWhitelist} from 'hooks/useDaoMembers';
-import {ActionParameter} from 'utils/types';
+import {useDaoActions} from 'hooks/useDaoActions';
 
 const NewProposal: React.FC = () => {
   const {data: dao, loading} = useDaoParam();
-  const {data: whitelist} = useDaoWhitelist(dao);
+  const {data: actions} = useDaoActions(dao);
   const [showTxModal, setShowTxModal] = useState(false);
 
   const {t} = useTranslation();
@@ -45,46 +44,6 @@ const NewProposal: React.FC = () => {
   if (loading) {
     return <Loading />;
   }
-
-  const baseActions: ActionParameter[] = [
-    {
-      type: 'withdraw_assets',
-      title: t('AddActionModal.withdrawAssets'),
-      subtitle: t('AddActionModal.withdrawAssetsSubtitle'),
-    },
-    {
-      type: 'modify_settings',
-      // TODO: Replace these with proper copies and i18n.
-      title: 'Modify Settings',
-      subtitle: 'Propose new settings for your DAO',
-    },
-    {
-      type: 'external_contract',
-      title: t('AddActionModal.externalContract'),
-      subtitle: t('AddActionModal.externalContractSubtitle'),
-    },
-  ];
-
-  const whitelistActions = baseActions.concat([
-    {
-      type: 'add_address',
-      title: t('AddActionModal.addAddresses'),
-      subtitle: t('AddActionModal.addAddressesSubtitle'),
-    },
-    {
-      type: 'remove_address',
-      title: t('AddActionModal.removeAddresses'),
-      subtitle: t('AddActionModal.removeAddressesSubtitle'),
-    },
-  ]);
-
-  const erc20Actions = baseActions.concat([
-    {
-      type: 'mint_token',
-      title: t('AddActionModal.mintTokens'),
-      subtitle: t('AddActionModal.mintTokensSubtitle'),
-    },
-  ]);
 
   return (
     <FormProvider {...formMethods}>
@@ -129,9 +88,7 @@ const NewProposal: React.FC = () => {
             </Step>
           </FullScreenStepper>
 
-          <AddActionMenu
-            actions={whitelist ? whitelistActions : erc20Actions}
-          />
+          <AddActionMenu actions={actions} />
         </ActionsProvider>
       </CreateProposalProvider>
     </FormProvider>
